--- a/src/libsync/CMakeLists.txt
+++ b/src/libsync/CMakeLists.txt
@@ -57,6 +57,7 @@ set(libsync_SRCS
     propagateremotedelete.cpp
     propagateremotemove.cpp
     propagateremotemkdir.cpp
+    propagatefiles.cpp
     syncengine.cpp
     syncfilestatus.cpp
     syncfilestatustracker.cpp