--- a/src/libsync/configfile.cpp
+++ b/src/libsync/configfile.cpp
@@ -52,6 +52,7 @@ static const char updateCheckIntervalC[] = "updateCheckInterval";
 static const char geometryC[] = "geometry";
 static const char timeoutC[] = "timeout";
 static const char chunkSizeC[] = "chunkSize";
+static const char smallFileSizeC[] = "smallFileSize";
 
 static const char proxyHostC[] = "Proxy/host";
 static const char proxyTypeC[] = "Proxy/type";
@@ -126,6 +127,12 @@ quint64 ConfigFile::chunkSize() const
 {
     QSettings settings(configFile(), QSettings::IniFormat);
     return settings.value(QLatin1String(chunkSizeC), 10*1000*1000).toLongLong(); // default to 10 MB
+}
+
+quint64 ConfigFile::smallFileSize() const
+{
+    QSettings settings(configFile(), QSettings::IniFormat);
+    return settings.value(QLatin1String(smallFileSizeC), 500*1000).toLongLong(); // default to 500 kB
 }
 
 void ConfigFile::setOptionalDesktopNotifications(bool show)