This fixes issues: 
#5391 In the mixed file size scenario bandwidth is very underutilized - enterprise setup
#5390 Broken sync time estimation - enterprise setup
#4498 sync files from smallest first to biggest last
#1633 Propagator: Balance concurrent up/down-load
#4454 Investigate HTTP Pipelining

This includes:
https://github.com/owncloud/client/pull/5400 so if this is merged, the other can be closed

This is basic for:
https://github.com/owncloud/client/issues/5319 Bundling
HTTP2

This can be enhanced by:
https://github.com/owncloud/client/pull/5368 Dynamic Chunking
https://github.com/owncloud/client/pull/5349 Sorting folders by modification time

This requires following capability on server side, otherwise will sync old way:
![selection_018](https://cloud.githubusercontent.com/assets/13368647/22213174/70dc713e-e193-11e6-98df-3ddd34f97f37.png)

### Idea

The idea behind the implemented algorithm is that in the scenario, in which we have a lot of mixed within the same or different folder - big files, under chunking size files of size 100kB or 5MB, files to download, moves, deletes, directory creations etc:

- Separate transfer jobs from jobs which server database interaction is major factor

- Reserve uncoditionaly 4-6 flows for "db sensitive files" like small file uploads and 2 for upload/download >1MB file transfers

- sync normalny if no other files

- Cross folder, this means it will look for "db sensitive" and "data transfer" in all folders.

The above will ensure that there is a balance between bandwidth utilization and bookkeeping, what in turn gives faster synchronisation (for details look below)

Used algorithm:
![selection_019](https://cloud.githubusercontent.com/assets/13368647/22213192/844d8c62-e193-11e6-8533-6584e281f523.png)

### Checkpoints

- [x] Cross folder, this means it will look for "db sensitive" and "data transfer" in all folders.
- [x] Test manualy to ensure faster sync - passed prove of concept
- [x ] Test on enterprise setup to ensure faster sync
- [x] Write unit tests to cover whole the code 
- [ ] Test with integration tests

### Details
Lets do some maths. This will be our folder to sync:
100 files - average 100kB file size -> total 10MB to be transfered
10 files - average 10MB file size -> total 100MB to be transfered

Assume, that your network is 5MB/s and one "1 Byte PUT" takes 1s (we are not in ideal home case scenario with empty server now guys, it is not that easy, it could take much much more)

The request "latency" consists of 2 components, time it takes for bookkeeping on server and time it takes for data transfer.

Current case

If you do 100 small request in the row, your data transfer is neglible ~0s and bookkeeping time is 100s. Parallelising that maybe you can achieve 33s having 3 parallel flows.

If you do 10 biger files request in a row, your data transfer is 10s bookkeeping. Parallelised say 4s if you are lucky with 3 parallel flows. However, you cannot omit 20s coming from transfer of 100MB, does not matter how many requests you have in parallel 1000 or 1. Your 5MB/s office net bounds you.

Total you need around 33s plus 24s from big, having 57s.

Optimized case

If you do 100 small request in the row, your data transfer is neglible ~0s and bookkeeping time is 100s. Parallelising that maybe you can achieve 50s reserving 2 flow slots. In this 50s, you used neglible bandwidth. If you use the 3rd flow to pump there 30s comming from big files, you just synced your files in 50s, since your 30s are done in parallel filling the bandwidth.

57s vs 50s is like 13% of the time orginal time? In this example it is around 7s. 100 files is not a big deal but shows bigger picture

More big files and more small files, the different is the percentage. Do the math for 55000 files and 55GB. Where you have 40000 of small files in 10GB(avg. size 250kB) and 15000 in the rest 45GB (avg size 3MB). I think we wont be talking about even minutes there :>

@felixboehm @DeepDiver1975 @davidjericho @cdamken @jnweiger