@phil-davis 2017-01-09
I made some minor edit suggestions for comments in #5441 
The principle of this seems very useful - for example I am a long way (in round-trip-time ping...) from my server and my upload speed is not so good (1 to 3 Mbps range). I often have cases where there are a few big files and a lot of small files. That often results in sub-optimal total transfer time, because a some big files might chug away for a while (3 in parallel), then it gets to a heap of small files and those then have to slowly get created on the server... If we make sure that the upload of lots of small files is running in one stream(s) in parallel with big files in another stream(s) then the things that need server time and the things that need raw transfer time can happen concurrently and the total elapsed time should be minimized.
👍 for the concept. I am not familiar enough with the overall client design to give a good opinion on the detailed implementation.