@mrow4a 2017-01-09
Ok, it works awesome, need to check it in enterprise scale. Remember, if you want to test it:

You need a server capability of enviromental capability on the client - otherwise it will sync the old way. 
```
    static const auto bundling = qgetenv("OWNCLOUD_BUNDLING");
    if (bundling == "0") return false;
    if (bundling == "1") return true;

    return _capabilities["dav"].toMap()["bundling"].toByteArray() >= "1.0";
```

@phil-davis Could you rebase and check again! I made a change to be sure that it parallelises correctly for ALL items. It was not previously, so 3 big downloads will block everything if placed in alphabetical order.