@mrow4a 2017-01-10
@ckamm Because PropagateDirectory is much more complicated thinking about cases to handle, did not want to introduce even more complexity, we will get lost. 

Also I wanted to make this "pluggable" so you can easily switch it on and off. 

Mind that PropagateFiles is only for new files without any "special" cases, just normal upload/download, chunked or not. No _first job, delete cases, WaitInDirectory etc.