@mrow4a 2017-01-23
Hello, 

Changed a little bit a structure of this and been testing it on enterprise setup, details can be found in this document @ogoffart @guruz @jturcotte @felixboehm, which is part of my CS3 presentation : 
SYNC OPTIMIZATIONS - SCHEDULING
[sync-optimizations-mrowczynski-scheduling.pdf](https://github.com/owncloud/client/files/724335/sync-optimizations-mrowczynski-scheduling.pdf)

Features:
- Ensured balance between jobs
- Is entriely plugable, cross-folder and separate from Propagate Directory main logic
- Only New and Updated files are being propagated (data transfers) 
- Jobs are created lazily