@guruz 2017-03-23
@mrow4a: make sure to try to interrupt syncs in the middle and then resume them (also multiple times) and check if this will really still lead to correctly synchronized directories and not to deletes file or files that are not synced.
@mrow4a: please make sure smashbox reliability sync tests pass
@mrow4a: please make sure S3 syncs faster than without the patch