@ogoffart 2017-04-28
I removed the milestone because this would still need more work, and I don't think it is worth it at all. But I leave it open for the sake of discussion.